+++ tests/Unit/Core/ConfigTest.php
@@ -228,30 +228,6 @@ class ConfigTest extends \OxidTestCase
         return $res;
     }
 
-    public function testIsUtfWhenInISOMode()
-    {
-        $config = $this->getMock(Config::class, array('getConfigParam'));
-        $config->expects($this->any())->method('getConfigParam')->with($this->equalTo('iUtfMode'))->will($this->returnValue(0));
-        $this->assertFalse($config->isUtf());
-    }
-
-    public function testIsUtfWhenInUtfMode()
-    {
-        $config = $this->getMock(Config::class, array('getConfigParam'));
-        $config->expects($this->any())->method('getConfigParam')->with($this->equalTo('iUtfMode'))->will($this->returnValue(1));
-        $this->assertTrue($config->isUtf());
-    }
-
-    /**
-     * This iUtfMode parameter was removed, as OXID eShop is UTF8 only since 6.0 version.
-     * So this method should return true.
-     */
-    public function testIsUtfWhenNoUtfModeDefined()
-    {
-        $config = oxNew('oxConfig');
-        $this->assertTrue($config->isUtf());
-    }
-
     private function _getOutPath($oConfig, $sTheme = null, $blAbsolute = true)
     {
         $sShop = $blAbsolute ? $oConfig->getConfigParam('sShopDir') : "";